First commit
This is the first commit of my laundry timer app. This is before I make it into an app. You can input a time in minutes that you want it to count down from and it will count down. You can add 30 minutes to the timer at any time and it will count down. You can reset the timer and you can also stop the timer