still trying to fix the inventory bug
I am trying to make it so that when I click the bag button it opens my inventory, I am then trying to make it so that a close bag button shows up in the inventory and I can click it and it will close the inventory and take me back to the original screen. still haven't figured it out. I have been using chatgpt but i think that it might be giving me wrong info. Will check online the next time I attempt this.